moving config flags to common package
